lambda/requirements_postgres.txt
Flask==2.3.3
Flask-CORS==4.0.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
shap==0.42.1
psycopg2-binary==2.9.9
joblib==1.3.2
boto3==1.34.0
